ForestControl-class {party}	R Documentation

Class "ForestControl"

Description

Objects of this class represent the hyper parameter setting for forest growing.

Objects from the Class

Objects can be created by cforest_control.

Slots

ntree:

number of trees in the forest.

replace:

sampling with or without replacement.

fraction:

fraction of observations to sample without replacement.

trace:

logical indicating if a progress bar shall be printed.

varctrl:

Object of class "VariableControl"

splitctrl:

Object of class "SplitControl"

gtctrl:

Object of class "GlobalTestControl"

tgctrl:

Object of class "TreeGrowControl"

Extends

Class "TreeControl", directly.

Methods

No methods defined with class "ForestControl" in the signature.

[Package party version 1.3-0 Index]