LearningSample Class {party}	R Documentation

Class "LearningSample"

Description

Objects of this class represent data for fitting tree-based models.

Objects from the Class

Objects can be created by calls of the form new("LearningSample", ...).

Slots

responses:

Object of class "VariableFrame" with the response variables.

inputs:

Object of class "VariableFrame" with the input variables.

weights:

Object of class "numeric", a vector of case counts or weights.

nobs:

Object of class "integer", the number of observations.

ninputs:

Object of class "integer", the number of input variables.

Methods

No methods defined with class "LearningSample" in the signature.

[Package party version 1.3-6 Index]